Add files via upload
no brand name or logo on heat pump.
g2 valve too big and the outflow does not line up with the supply that goes to the DHW
can we control more features of the flow dots?  I think the white part of the dot is too big(stroke width?) and would like to control it. lets add a configuration option
instead of only cycling between the words HEAT/DHW when the G2 mode switches lets also animate the valve.  Maybe use the red filled sections of the valve from my illustration?

Let's add the cylinder around the supply pipe between the hp and the g2 for the v18.  It's presence needs to be triggered by settings in the configuration file.  something that wraps around the pipe and is semi-transparent could work well.  It can even look like a coil if you can draw that.  The coil can glow rad when it is enabled, and we can control the intensity of the glow based on the power going to the v18.  The intensity could become visible by a change in color combined with a change in intensity of the shadow of the coil.  Do you think that would work visually?